refactor: 채널 코드 리팩토링

+++ srcs/requirements/backend/nest/src/chat/service/channel/mute.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { Cron } from "@nestjs/schedule";
 import { InjectRepository } from "@nestjs/typeorm";
+import { WsException } from "@nestjs/websockets";
 import { Mute } from "src/chat/entity/channel/mute.entity";
 import { LessThanOrEqual, Repository } from "typeorm";
 
@@ -11,19 +12,19 @@ export class MuteService {
     private readonly muteRepository: Repository<Mute>
   ) {}
 
-  async has(channelId: number, userId: number): Promise<boolean> {
+  async verify(channelId: number, userId: number): Promise<void> {
     const primaryKey = {
       channelId,
       userId,
     };
-    try {
-      const mute: Mute = await this.muteRepository.findOneByOrFail(primaryKey);
-      if (mute.limitedAt > new Date()) {
-        return true;
-      }
-      await this.muteRepository.delete(primaryKey);
-    } catch {}
-    return false;
+    const mute: Mute = await this.muteRepository.findOneBy(primaryKey);
+    if (!mute) {
+      return;
+    }
+    if (mute.limitedAt > new Date()) {
+      throw new WsException("can't send because muted user.");
+    }
+    await this.muteRepository.delete(primaryKey);
   }
 
   @Cron("0 * * * *")
